+++ CHANGELOG.md
@@ -1,4 +1,8 @@
+TODO:
+ - Implement required filters, that is filters without which the URL is invalid.
 
+## 0.1.2 (2026-01-01)
+- Started work the Dart code generation.  Progress on enums. 
 
 
 ## 0.1.1 (2025-12-30)